New element_with_? convenience features
These features allow the creation of new elements with a single line rather than creation followed by attribute setting. This is a very constant process in creating the web page HTML -- that is -- we want to have a single line command with arguments, which represent the settings of the attributes so the entire creation exists on a single line.